+++ py/autoevals/oai.py
@@ -2,11 +2,13 @@ import asyncio
 import json
 import os
 import sqlite3
+import sys
+import textwrap
 import threading
 import time
 from pathlib import Path
 
-from .util import current_span, prepare_openai_complete
+from .util import current_span
 
 _CACHE_DIR = None
 _CONN = None
@@ -31,6 +33,56 @@ def open_cache():
 
 
 CACHE_LOCK = threading.Lock()
+
+
+def prepare_openai_complete(is_async=False, api_key=None):
+    try:
+        import openai
+    except Exception as e:
+        print(
+            textwrap.dedent(
+                f"""\
+            Unable to import openai: {e}
+
+            Please install it, e.g. with
+
+              pip install 'openai'
+            """
+            ),
+            file=sys.stderr,
+        )
+        raise
+
+    openai_obj = openai
+    is_v1 = False
+    if hasattr(openai, "OpenAI"):
+        # This is the new v1 API
+        is_v1 = True
+        if is_async:
+            openai_obj = openai.AsyncOpenAI(api_key=api_key)
+        else:
+            openai_obj = openai.OpenAI(api_key=api_key)
+
+    try:
+        from braintrust.oai import wrap_openai
+
+        openai_obj = wrap_openai(openai_obj)
+    except ImportError:
+        pass
+
+    complete_fn = None
+    rate_limit_error = None
+    if is_v1:
+        rate_limit_error = openai.RateLimitError
+        complete_fn = openai_obj.chat.completions.create
+    else:
+        rate_limit_error = openai.error.RateLimitError
+        if is_async:
+            complete_fn = openai_obj.ChatCompletion.acreate
+        else:
+            complete_fn = openai_obj.ChatCompletion.create
+
+    return complete_fn, rate_limit_error
 
 
 def post_process_response(resp):
